Fix compilation when bytemuck is used with the derive feature
With the feature enabled there's also a bytemuck::Pod derive macro,
which will conflict with mango_macro::Pod.

Using bytemuck's Pod macro instead would be a larger change because
it validates that all fields in the struct are Pod themselves. And
technically neither bool or enums are: these types do not allow any
bit pattern.